+++ hpg-bigdata-app/src/main/java/org/opencb/hpg/bigdata/app/cli/executors/AlignmentCommandExecutor.java
@@ -358,7 +358,7 @@ public class AlignmentCommandExecutor extends CommandExecutor {
         AlignmentOptions options = new AlignmentOptions();
         options.setContained(false);
 
-        short[] values;
+        float[] values;
 
         BamManager alignmentManager = new BamManager(filePath);
         Iterator<SAMSequenceRecord> iterator = fileHeader.getSequenceDictionary().getSequences().iterator();
